Capture Modbus snapshot.  Not working.
