+++ CMakeLists.txt
@@ -245,8 +245,13 @@ set ( libplctag_SRCS "${lib_SRC_PATH}/init.c"
                      "${ab_SRC_PATH}/session.c"
                      "${ab_SRC_PATH}/session.h"
                      "${ab_SRC_PATH}/tag.h"
-                     "${mb_SRC_PATH}/modbus.c"
+                     "${mb_SRC_PATH}/mb_common.h"
+                     "${mb_SRC_PATH}/mb_common.c"
                      "${mb_SRC_PATH}/modbus.h"
+                     "${mb_SRC_PATH}/plc.c"
+                     "${mb_SRC_PATH}/plc.h"
+                     "${mb_SRC_PATH}/tag.c"
+                     "${mb_SRC_PATH}/tag.h"
                      "${protocol_SRC_PATH}/system/system.c"
                      "${protocol_SRC_PATH}/system/system.h"
                      "${protocol_SRC_PATH}/system/tag.h"
